requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pyyaml==6.0.2
rapidfuzz==3.11.0
apscheduler==3.10.4
aiofiles==24.1.0
